Fixed URL-creation bug in AddOrEditHyperlink script
Now it won't create shared hyperlinks by default.

This means that if you copy and paste text with a hyperlink in it that was
created by this script, the script will create a new hyperlink object along
with the new text object, rather than just copying over a reference
to the old hyperlink object as it did before.